// server/intelligence-service/src/db/schema.ts
export const user = pgTable("user", {
	// You can use { mode: "bigint" } if numbers are exceeding js number limitations
	id: bigint({ mode: "number" }).primaryKey().notNull(),
	createdAt: timestamp("created_at", { withTimezone: true, mode: "string" }),
	updatedAt: timestamp("updated_at", { withTimezone: true, mode: "string" }),
	avatarUrl: varchar("avatar_url", { length: 255 }),
	blog: varchar({ length: 255 }),
	company: varchar({ length: 255 }),
	description: varchar({ length: 255 }),
	email: varchar({ length: 255 }),
	followers: integer().notNull(),
	following: integer().notNull(),
	htmlUrl: varchar("html_url", { length: 255 }),
	location: varchar({ length: 255 }),
	login: varchar({ length: 255 }),
	name: varchar({ length: 255 }),
	type: varchar({ length: 255 }),
	leaguePoints: integer("league_points").default(0).notNull(),
	notificationsEnabled: boolean("notifications_enabled")
		.default(true)
		.notNull(),
});
// ...
export const document = pgTable(
	"document",
	{
		id: uuid().notNull(),
		versionNumber: integer("version_number").notNull(),
		content: text(),
		createdAt: timestamp("created_at", {
			precision: 6,
			withTimezone: true,
			mode: "string",
		}).notNull(),
		kind: varchar({ length: 255 }).notNull(),
		title: varchar({ length: 255 }).notNull(),
		// You can use { mode: "bigint" } if numbers are exceeding js number limitations
		userId: bigint("user_id", { mode: "number" }).notNull(),
	},
	(table) => [
		index("idx_document_created_at").using(
			"btree",
			table.createdAt.asc().nullsLast().op("timestamptz_ops"),
		),
		index("idx_document_id").using(
			"btree",
			table.id.asc().nullsLast().op("uuid_ops"),
		),
		index("idx_document_user_id").using(
			"btree",
			table.userId.asc().nullsLast().op("int8_ops"),
		),
		foreignKey({
			columns: [table.userId],
			foreignColumns: [user.id],
			name: "FKjhdxdv9sijhujiynqbb5jc010",
		}),
		primaryKey({
			columns: [table.id, table.versionNumber],
			name: "documentPK",
		}),
	],
);